requirements.txt
crewai>=0.51.1
crewai-tools>=0.8.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
requests>=2.32.0
pydantic>=2.7.0
pandas>=2.2.0
fpdf2>=2.7.8
streamlit>=1.38.0
python-dotenv>=1.0.1
langchain>=0.3.1
langchain-community>=0.3.1
langchain-google-genai>=2.0.0
chromadb>=0.5.4
PyMuPDF>=1.24.8
